backend/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
SQLAlchemy==2.0.36
asyncpg==0.29.0
python-dotenv==1.0.1
psycopg-binary==3.2.3
pgvector==0.3.3
httpx==0.27.2
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.10
email-validator==2.2.0
alembic==1.17.2
redis==5.0.8
rq==1.16.2
